+++ CHANGELOG.md
@@ -9,3 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [dependabot]: Bump actions/checkout from 4 to 7 in the all-deps group ([#209](../../pull/209))
+- [BUGFIX] Update validate-pr workflow to work with dependabot PRs (#0) ([#211](../../pull/211))
+- [HOTFIX] Changelog workflow now pushes commit to branch with PR  (#0) ([#210](../../pull/210))